pilot_assist/requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sentence-transformers>=2.2.2
tiktoken>=0.5.0 